% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetric_matrix.R
\name{calculate_maximum}
\alias{calculate_maximum}
\title{List of profiles to calculate the maximum average value of variable across
all profiles in the list.}
\usage{
calculate_maximum(profiles)
}
\arguments{
\item{profiles}{A list of data frames or matrices}
}
\description{
List of profiles to calculate the maximum average value of variable across
all profiles in the list.
}
