% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pineplots.R
\name{draw_heatmap}
\alias{draw_heatmap}
\title{Plots a heatmap of a symmetric matrix.}
\usage{
draw_heatmap(sym_matrix, fill_name = "Value", xlabel = "",
  ylabel = "", limits = NULL, pyramid = TRUE, low = "blue",
  mid = "white", high = "red", midpoint = 0, breaks = waiver())
}
\arguments{
\item{sym_matrix}{A matrix of symmetric values indicating a relationship between variables of the rows/columns.}

\item{fill_name}{Name of values in the symmetric matrix.}

\item{xlabel}{Label to include on x-axis.}

\item{ylabel}{Label to include on y-axis.}

\item{limits}{Limits of the fill values.}

\item{pyramid}{Boolean indicates lower triangle of the heamap be removed (TRUE) or not (FALSE).}

\item{low}{Lower colour for fill gradient.}

\item{mid}{Midpoint colour for fill gradient.}

\item{high}{Higher colour for fill gradient.}

\item{midpoint}{The midpoint of the legend scale (if legend included).}

\item{breaks}{Position of the tick marks in the heatmap legend.}
}
\value{
A ggplot heatmap object that can be further manipulated using ggplot components if necessary.
}
\description{
Plots a heatmap of a symmetric matrix.
}
