% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{heart}
\alias{heart}
\title{Cleveland heart disease dataset}
\format{An object of class \code{data.frame} with 303 rows and 14 columns.}
\usage{
heart
}
\description{
A description.
}
\keyword{datasets}
