% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{layers}
\alias{layers}
\title{Lists of example symmetric matrices.}
\format{Contains 3 lists with:
\describe{
  \item{layers_B}{symmetric matrices for brain tissue}
  \item{layers_K}{symmetric matrices for kidney tissue}
  \item{layers_L}{symmetric matrices for liver tissue}
  ...
}}
\usage{
layers
}
\description{
A dataset containing the mutual expression values between
15 genes that are either predicted as liver or kidney specific.
Each species mouse, chicken, and macaque have a symmetric matrix
stored in each list contain in layers.
}
\keyword{datasets}
