% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{merkin_brawand_rnaseq}
\alias{merkin_brawand_rnaseq}
\title{Mergin and Brawand gene expression dataset}
\format{An object of class \code{matrix} with 15 rows and 37 columns.}
\usage{
merkin_brawand_rnaseq
}
\description{
A description.
}
\keyword{datasets}
