% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pineplot.R
\docType{package}
\name{pineplot}
\alias{pineplot}
\alias{pineplot-package}
\title{pineplot: A package for generating ggplot2 triangular heatmaps and stacked triagular
heatmaps from symmetric matrices.}
\description{
The pineplot package provides functions to address the following objectives:
build symmetric matrices, visualize as ggplot2 triangular heat maps and stack them
for easy comparison across group, time, phenotype etc. The package allows for
easy manipulation of heat maps before pine plot construction.
}
\section{make_pineplots functions}{

The make_pineplots functions
}

\section{symmetric_matrix functions}{

These functions provide an example of one way to construct a symmetric matrix in R and
measure a relationship between variables of interest. The symmetric_matrix_genorator
function
}

