% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetric_matrix.R
\name{symmetric_matrix_generator}
\alias{symmetric_matrix_generator}
\title{Function that calculates the mutual expression between genes in a expression table.}
\usage{
symmetric_matrix_generator(exp_matrix, max.expression)
}
\arguments{
\item{exp_matrix}{The expression matrix where rows are the genes
and columns are the samples/observations.}

\item{max.expression}{is the maximum average value to normalize the average values
of variables in exp_matrix.}
}
\value{
A symmetric matrix, \code{sym_mat}, that contains mutual expression value
between 0 and 1 to symbolize relationship between variable x and variable y.
}
\description{
Function that calculates the mutual expression between genes in a expression table.
}
