% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tissue_specific_genes}
\alias{tissue_specific_genes}
\title{Log2 gene expression data for kidney and liver "specific" genes.}
\format{A matrix with 15 rows and 37 variables. The rows of the matrix
represent genes that are predicted as being liver (first 10 rows) specific
or kidney (last 5 rows) specific.
Mouse samples are stored in columns 1-12, chicken in columns 13-24, and
macaque in columns 25-37.
The 37 variables are labelled as follows:
\describe{
  \item{B}{Brain samples}
  \item{K}{Kidney samples}
  \item{L}{Liver samples}
}}
\usage{
tissue_specific_genes
}
\description{
A dataset containing the normalized gene expression data from
mouse, chicken, and macaque RNA-seq data.
}
\examples{
data(tissue_specific_genes)
ch_grp <- c("B", "B", "K", "K", "L", "L", "B", "K", "B", "K", "L", "K")
mu_grp <- c("B", "B", "K", "K", "L", "L", "B", "K", "L", "B", "K", "L")
maca_grp <- c("B", "B", "K", "K", "L", "L", "L", "B", "K", "L", "B", "K", "L")
mouse <- 1:12
chicken <- 13:24
macaque <- 25:37

chicken.profile <- tissue_specific_genes[, chicken][, which(ch_grp == "L")]
mouse.profile <- tissue_specific_genes[, mouse][, which(mu_grp == "L")]
macaque.profile <- tissue_specific_genes[, macaque][, which(maca_grp == "L")]
}
\keyword{datasets}
