% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pineplots.R
\name{write_pine_plot}
\alias{write_pine_plot}
\title{Make the pine plot for a list of heatmaps and organize in a stacked, pine tree shape.}
\usage{
write_pine_plot(h_maps, num_heatmaps = length(h_maps), height = 30,
  width = 10, leg = FALSE, angle = c(-45), legend_x_offset = 0.02,
  legend_y_offset = 0, hm_y_offset = 0, hm_margin = 0)
}
\description{
@param h_maps List of heat maps to add to a pineplot.
 @param num_heatmaps The total number of heat maps to add to pine plot.
 @param height The height of the file to start plotting the heat maps.
 @param leg Boolean indicating if plots have a legend.
 @param angle angle to rotate each heat map.
 @param legend_x_offset Depending on label size and length, may want to adjust position of the legend using offset.
 @param legend_y_offset Depending on label size and length, may want to adjust the y position of the pine plot.
 @param hm_y_offset Changes the y posiiton of heatmap shifting them higher or lower in the viewport.
 @param hm_margin Changes the space left between the heat maps if they need to be closer or farther appart for athstetic purpose.
 @return Pine plot written to view or a file if open. No return object.
 @export
}
