% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pineplots.R
\name{draw_heatmap}
\alias{draw_heatmap}
\title{Plots a heatmap of a symmetric matrix.}
\usage{
draw_heatmap(sym.matrix, fill.name = "Value", xlabel = "",
  ylabel = "", limits = NULL, pyramid = TRUE, breaks = waiver(),
  midpoint = 0, barwidth = 5, barheight = 1,
  direction = "horizontal", legend = TRUE, low = "blue",
  mid = "white", high = "red")
}
\arguments{
\item{sym.matrix}{A matrix of symmetric values indicating a relationship between variables of the rows/columns.}

\item{fill.name}{Name of values in the symmetric matrix.}

\item{xlabel}{Label to include on x-axis.}

\item{ylabel}{Label to include on y-axis.}

\item{limits}{Limits of the fill values.}

\item{pyramid}{Boolean indicates lower triangle of the heamap be removed (TRUE) or not (FALSE).}

\item{breaks}{Number of break points in the legend (if legend included).}

\item{midpoint}{The midpoint of the legend scale (if legend included).}

\item{barwidth}{Width of the legend (if legend included).}

\item{barheight}{Height of the legend (if legend included).}

\item{direction}{Legend direction either "vertical" or "horizontal".}

\item{legend}{Boolean value indicating whether or not to include a legend for the plot.}

\item{low}{Lower colour for fill gradient.}

\item{mid}{Midpoint colour for fill gradient.}

\item{high}{Higher colour for fill gradient.}
}
\value{
A ggplot heatmap object that can be further manipulated using ggplot components if necessary.
}
\description{
Plots a heatmap of a symmetric matrix.
}
