% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pineplots.R
\name{draw_heatmap}
\alias{draw_heatmap}
\title{Plots a heatmap of a symmetric matrix.}
\usage{
draw_heatmap(sym.matrix, fill.name = "Value", xlabel = "",
  ylabel = "", lim = NULL, pyramid = TRUE, breaks = waiver(),
  midpoint = 0, barwidth = 5, barheight = 1, text_size = 8,
  direction = "horizontal", legend = TRUE, low = "blue",
  mid = "white", high = "red")
}
\description{
@param sym.matrix A matrix of symmetric values indicating a relationship between variables of the rows/columns.
 @param fill.name Name of values in the symmetric matrix.
 @param xlabel Label to include on x-axis.
 @param ylabel Label to include on y-axis.
 @param lim Limits of the fill values.
 @param pyramid Boolean indicates lower triangle of the heamap be removed (TRUE) or not (FALSE).
 @param breaks Number of break points in the legend (if legend included).
 @param midpoint The midpoint of the legend scale (if legend included).
 @param barwidth Width of the legend (if legend included).
 @param barheight Height of the legend (if legend included).
 @param direction Legend direction either "vertical" or "horizontal".
 @param legend Boolean value indicating whether or not to include a legend for the plot.
 @return A ggplot heatmap object that can be further manipulated using ggplot components if necessary.
 @examples
 draw_heatmap(matrix(c(2,1,0,3,0,1,0,3,2),3,3), lim=c(0,3),breaks=c(0,1.5,3),midpoint=1.5)
 draw_heatmap(matrix(c(-0.5,1,0,0.75,0,1,0,0.75,-0.5),3,3))
 @export
}
