% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pineplots.R
\name{generate_pineplot}
\alias{generate_pineplot}
\title{Make a pine plot for a list of symmetric matrices.}
\usage{
generate_pineplot(sym_matrices, annotation_fn, customize_fn, scale = 1,
  legend_scale = 1, legend = TRUE, ...)
}
\arguments{
\item{sym_matrices}{List of heat maps to add to a pineplot.}

\item{annotation_fn}{Annotation function for ggplot heatmaps.}

\item{customize_fn}{Customization function for the heatmap viewports.}

\item{scale}{Scaling coefficient applied to each heatmap. Necessary in some instances to prevent overlap or reduce the whitespace between plots.}

\item{legend_scale}{Scaling factor for legend size.}

\item{legend}{Logical value indicating whether to include a colourbar legend or not.}

\item{...}{Additional arguments to be passed to \code{\link{draw_heatmap}}.}
}
\value{
No value is returned. The plot is drawn onto the current device or viewport.
}
\description{
Make a pine plot for a list of symmetric matrices.
}
